Cheetah/requirement.txt
decord==0.6.0
gradio==3.39.0
iopath==0.1.10
matplotlib==3.7.1
numpy==1.24.3
omegaconf==2.3.0
opencv_python_headless==4.5.5.64
pandas==2.0.2
Pillow==10.0.0
PyYAML==6.0
Requests==2.31.0
ruamel.base==1.0.0
scikit_learn==1.3.0
scipy==1.10.1
timm==0.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
SentencePiece==0.1.99
rouge==1.0.1